+++ lib/code_lister/code_lister.rb
@@ -1,18 +1,52 @@
 module CodeLister
   CustomError = Class.new(StandardError)
   class << self
-    # List files base on some extension
+    # Execute the command in the shell and extract the output to be used
+    # e.g. `git diff --name-only HEAD~1` is getting the list of files that have been
+    # updated in the last commit
+    #
+    # @param [String] command the input command to be executed in the shell
+    # @param [String] base_dir the starting directory
+    # @return [Array<String>] file list or empty list if the shell command is not valid
+    def files_from_command(command, base_dir)
+      files = AgileUtils::Helper.shell(command.split(" ")).split(/\n/)
+      # Adapt the result and make sure that it starts with "./"
+      # like the result from 'CodeLister.files()' method
+      files.map! do |file|
+        if file =~ /^\.\//
+          # skip if the file start with './' string
+          file
+        else
+          # add './' to the one that does not already have one
+          "./#{file}"
+        end
+      end
+      # Note: this make sure that it works with deleted file when use
+      # this with something like 'git diff --name-only HEAD~2'
+      files.delete_if { |file| !File.exist?(file.gsub(/^\./, base_dir)) }
+    rescue RuntimeError => e
+      # just return the empty list, if the user specified invalid command
+      return []
+    end
+
+    # List files base on multiple simple criteria
+    #
+    # @param [Hash<Symbol>,<Object>] args argument hash
+    # @option args [String]         :base_dir the starting directory
+    # @option args [Boolean]        :recursive flag to indicate if the search will be done recursively
+    # @option args [Array<String>]  :exts list of file extension to search for (without the dot)
+    # @option args [Array<String>]  :non_exts list of file without any extension to search for
+    # @return [Array<String>] the list of file based on the matching criteria
     def files(args = {})
       opts = {
-        base_dir: Dir.pwd,
+        base_dir:  Dir.pwd,
         recursive: false,
-        exts: [],
-        non_exts: []
+        exts:      [],
+        non_exts:  []
       }.merge(args)
 
       # always expand the path
       base_dir = File.expand_path(opts[:base_dir])
-      # base_dir = opts[:base_dir]
       fail CustomError, "The directory #{base_dir} is not valid or or not readable!" unless File.exist?(base_dir)
 
       wildcard = opts[:recursive] ? "**" : ""
@@ -28,7 +62,14 @@ module CodeLister
       (files_with_extension + files_without_extension).sort
     end
 
-    # Filter out the list based on given list of words
+    # Filter out the list based on simple criteria
+    #
+    # @param [Array<String>] file_list the input file list
+    # @param [Hash<Symbol, Object>] args the option hash
+    # @option args [Array<String>]  :inc_words list of words that must be match to be included in the result
+    # @option args [Array<String>]  :exc_words list of words that will be excluded fromt the result
+    # @option args [Boolean]        :ignore_case flag to indicate how the string comparision should be performed
+    # @return [Array<String>] the original list with the result filtered
     def filter(file_list, args = {})
       opts = {
         inc_words: [],